Week 5 scripts complete
This commit includes a modification to a Week 4 script due to a command executed
in a Week 5 script.
It should be noted that there is one regex command in regexes.sh that doesn't
seem to return any values. After doing A LOT of research, I see no reason as
to why it doesn't work, so I'm leaving it in until I can determine a better
replacement.